Use min/max from stdlib
Until now, tools.h had it's own definition of the min and
max functions. We replace those with the ones from the
stdlib.

In order for this to work we remove the functions from
tools.h and replace all the invocations with ones that use
the std namespace `std::max(a,b)`.

This was done mostly via automatic replacement with few
minor manual adjustments. It would not be feasable to do
that manually, but the semi automatic approach might have
created some artifacts.

== minmax.sh ===============================================

 # ag is a more clever grep
ag 'max|min' -l | xargs sed -i '
  s@[-]max[(]@-std::max(@g;
  s@[+]max[(]@+std::max(@g;
  s@[*]max[(]@*std::max(@g;
  s@[/]max[(]@/std::max(@g;
  s@[=]max[(]@=std::max(@g;
  s@[(]max[(]@(std::max(@g;
  s@[)]max[(]@)std::max(@g;
  s@[[]max[(]@[std::max(@g;
  s@ max[(]@ std::max(@g;

  s@[-]::max[(]@-std::max(@g;
  s@[+]::max[(]@+std::max(@g;
  s@[*]::max[(]@*std::max(@g;
  s@[/]::max[(]@/std::max(@g;
  s@[=]::max[(]@=std::max(@g;
  s@[(]::max[(]@(std::max(@g;
  s@[)]::max[(]@)std::max(@g;
  s@ ::max[(]@ std::max(@g;

  s@[-]min[(]@-std::min(@g;
  s@[+]min[(]@+std::min(@g;
  s@[*]min[(]@*std::min(@g;
  s@[/]min[(]@/std::min(@g;
  s@[=]min[(]@=std::min(@g;
  s@[(]min[(]@(std::min(@g;
  s@[)]min[(]@)std::min(@g;
  s@[[]min[(]@[std::min(@g;
  s@ min[(]@ std::min(@g;

  s@[-]::min[(]@-std::min(@g;
  s@[+]::min[(]@+std::min(@g;
  s@[*]::min[(]@*std::min(@g;
  s@[/]::min[(]@/std::min(@g;
  s@[=]::min[(]@=std::min(@g;
  s@[(]::min[(]@(std::min(@g;
  s@[)]::min[(]@)std::min(@g;
  s@ ::min[(]@ std::min(@g;
'